+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **BREAKING**: Use `artisan about` instead of `customApplicationData`.
 
 ## [0.2.2] - 2022-12-27
 ### Added